requirements.txt

streamlit
qrcode
pillow
